Original change: https://android-review.googlesource.com/c/platform/frameworks/support/+/3823489

+++ compose/remote/remote-creation-compose/src/main/java/androidx/compose/remote/creation/compose/state/RemoteLong.kt
@@ -145,7 +145,7 @@ public fun rememberRemoteLongValue(
         val initial = value()
         MutableRemoteLong(mutableLongStateOf(initial), constantValue = null) { creationState ->
             val id = creationState.document.addNamedLong(name, initial)
-            creationState.document.setStringName(id.toInt(), "$domain:$name")
+            creationState.document.setStringName(id, "$domain:$name")
             id
         }
     }
